requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==3.2
django-babel==0.6.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-phonenumber-field==7.0.0
django-psycopg2-extension==0.1.1
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2==2.7.5
pytz==2022.5
sqlparse==0.4.3
typing-extensions==4.4.0
